'Refactor: update imports to use specific path for post model in various modules'
This commit refactors the imports of the Post model in various modules to use a specific path for the post model. Specifically, the import statement for
