+++ src/app.module.ts
@@ -8,7 +8,7 @@ import { Role } from './roles/model/roles.model'
 import { UserRoles } from './roles/model/user-roles.model'
 import { AuthModule } from './auth/auth.module'
 import { PostsModule } from './posts/posts.module'
-import { Post } from './posts/posts.model'
+import { Post } from './posts/model/posts.model'
 import { FilesModule } from './files/files.module'
 import { ServeStaticModule } from '@nestjs/serve-static'
 import * as path from 'path'
