+++ src/pipes/validation.pipe.ts
@@ -10,7 +10,7 @@ export class ValidationPipe implements PipeTransform<any> {
         const errors = await validate(obj)
 
         if (errors.length) {
-            let messages = errors.map((err) => {
+            const messages = errors.map((err) => {
                 return {
                     [err.property]: Object.values(err.constraints).join(', '),
                 }
