+++ src/posts/model/posts.model.ts
