+++ src/posts/posts.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common'
 import { PostsService } from './posts.service'
 import { PostsController } from './posts.controller'
-import { Post } from './posts.model'
+import { Post } from './model/posts.model'
 import { SequelizeModule } from '@nestjs/sequelize'
 import { User } from 'src/users/model/users.model'
 import { FilesModule } from 'src/files/files.module'
