+++ src/users/model/users.model.ts
@@ -8,7 +8,7 @@ import {
     Model,
     Table,
 } from 'sequelize-typescript'
-import { Post } from 'src/posts/posts.model'
+import { Post } from 'src/posts/model/posts.model'
 import { Role } from 'src/roles/model/roles.model'
 import { UserRoles } from 'src/roles/model/user-roles.model'
