+++ src/users/users.module.ts
@@ -8,7 +8,7 @@ import { RolesModule } from 'src/roles/roles.module'
 import { AuthService } from 'src/auth/auth.service'
 import { AuthModule } from 'src/auth/auth.module'
 import { UserRoles } from 'src/roles/model/user-roles.model'
-import { Post } from 'src/posts/posts.model'
+import { Post } from 'src/posts/model/posts.model'
 
 @Module({
     controllers: [UsersController],
